Move point cloud display files to rviz_common
